scripts/requirements.txt
anthropic>=0.25.0
feedparser>=6.0.11
requests>=2.31.0
supabase>=2.0.0
python-dotenv>=1.0.0
